iod: stub in handling of BYE message
Set the peer's done flag. If the number of completed requests is the
same as the number of requests recevied, we have serviced all requests
and can print results. Still need to handle the case when they are not
equal.

Still need to actually print the results to the file.