Change writeThread to do as much work as possible every time it wakes
Change the writeThread to process all the blocks in the ready list each
time it wakes (instead of just one).  Add some comments about when it's
safe to delete Peer objects from peerList.  Add some code to ensure
we're not accepting messages from a connection once we receive the
FINISHED message.