Daemon can report its cache usage
Add a pair of CCI messages so the clients can query the daemon's cache
usage and the daemon can reply.  Currently only replying with a single
bool indicating the cache is empty.

Changed how the -s option on the client works:  The default is now -1,
which means "sleep until the cache is drained".  It uses the previously
mentioned CCI messages to find out if the cache is drained.  (Positive
values for -s still cause it to sleep for a fixed number of seconds.)