Fix bug with iterator handling in peerList
map::erase() invalidates the iterator.  Had to change the code so that I
wasn't trying to increment said invalid iterator.  Frustratingly, the
code worked fine on my PC; it only failed on Titan...

Also got a little more stringent about locking the peerList mutex.  It
wasn't the cause of the problem, but it's probably a good idea.