Several minor fixes
1) Forgot to initialize daemonPid to -1 (which meant it was 0 for ranks
that didn't start the daemon.  Those ranks ended up calling waitPid()
when they didn't need to.)

2) Moved the call to startOneDaemon up before the call to MPI_Init().
(Calling fork() after MPI has started is aparently not something one
should do.  In our case, it seemed to sometimes cause the parent to
receive a strange signal when the child exited.)

3) Add leading 0's to the rank part of the filenames so they sort
correctly in a dir listing

4) Print a few debug messages during client shutdown.