Partial support for blocking mode
Added code to enable blocking mode on the daemon.  Also have partial
support for blocking mode on the client.  (There's still a little more
work to do to enable it.)

THIS CODE IS UNTESTED!!  The version of CCI that I have doesn't support
blocking mode for the shared memory transport.