Change how the -s parameter works
If -s is positive, then it's the same as before: sleep for a fixed
amount of time.  If -s is negative, then the code waits until the daemon
reports that the cache is empty.  It then calls sync() and waits the
specified amount of time (presumably to let the OS page cache drain).